+++ packages/devtools_app_shared/lib/src/ui/split_pane.dart
